Remove outdated conda libraries
Conda is shipped with RPATH and obsolete libraries. When opening a
shared object needing a library by that name, the conda lib takes over,
and fails miserably due to it missing a symbol. We *cannot* go over a
RPATH without preload, so we apply a little fix by erasing those
libraries !